pyproject.toml
[project]
name = "stock-advisor"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bs4>=0.0.2",
    "pandas>=2.2.3",
    "plotly>=6.0.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "streamlit>=1.44.0",
    "yfinance>=0.2.55",
    "google-cloud-storage>=2.14.0",
]
